feat: add TagHandler to validate VM tags
TagHandler has been added to the Common directory in AzureChronos.Functions to help with the validation of tags associated with VirtualMachineResource objects. The newly added class has a ValidateVirtualMachineTag method, which checks if a tag is associated with a given VM resource. This was implemented to enhance resource management by ensuring that only VMs with specific tags are processed.