feat: add StringHandler class to extract resource group name
This commit introduces a new utility class, `StringHandler`, in the `AzureChronos.Functions.Common` namespace. The purpose of this class is to provide helper methods for handling strings. Specifically, a method `ExtractResourceGroupName` has been added which, given an Azure resource ID, extracts and returns the name of the resource group. This method will enhance code readability and reusability when working with Azure resource IDs.