feat: implemented Azure Compute Service to Timer Trigger
Added Azure Compute Service to Azure Chronos Timer Trigger. This change allows for more granular control and querying of Azure virtual machines. The code now enumerates all VMs in the subscription, skips those for which an event has already been scheduled, and initializes an entity for new ones. It's aimed to enhance the management and scalability of Chronos service by allowing interactions with VMs on a subscription basis.