feat: add eligibility validation for virtual machines
A new method, ValidateVirtualMachineEligibilityAsync(), was added to the IVirtualMachineEntity interface and the VirtualMachineEntity class. This is to provide a way to validate if a virtual machine is eligible to be scheduled, which is critical for managing resources effectively in the system. Currently, the function in the class throws a NotImplementedException, actual functionality to be implemented in subsequent tasks.