feat: add function to fetch specific Azure VM
A new method, GetAzureVirtualMachineAsync, has been added to the AzureComputeService and its interface, IAzureComputeService. This function takes in subscriptionId, resourceGroupName and vmName as parameters and returns data on a specific Azure Virtual Machine. This feature was added to facilitate fetching specific VMs from Azure, instead of listing all of them every time. A Model Namespace was also added for handling the Virtual Machine Resource.