feat: add logic to validate VM eligibility
Updated the method ValidateVirtualMachineEligibilityAsync in the VirtualMachineEntity class to interact with the Azure API and check if a virtual machine meets the criteria for scheduling based on the "AzChronos_Startup" tag. Also added an import statement for the common functions. The changes aim to improve the scheduling process and incorporate a dynamic approach to validate the virtual machines. The application will now use actual data from Azure to determine the eligibility of each virtual machine.