feaet: add Azure Chronos scheduling tags and validation
Added functionality for recognizing specific Azure Chronos tags and validating if they are present in the virtual machine. The tags we are looking for are "AzChronos_Exclude", "AzChronos_Deallocate", and "AzChronosStartup". These tags will be used to determine if a VM is eligible for scheduling.

Further, code to retrieve and process schedule from the VM's tags was implemented and a new model 'AzureChronosSchedule' was added to contain this schedule information.

Also, a redundant method remark and example in AzureComputeService.cs has been removed to improve code readability.

This modification allows us to have a clearer understanding of the state and scheduling of each VM and presents a way to govern VM scheduling directly through the VM's tags.