feat: add essential packages and setup GRPC support to fix an issue on Apple Silicon devices.
This commit adds several important package references to the AzureChronos.Isolated project, including GRPC Core library (Contrib.Grpc.Core.M1), Azure Functions Worker Extensions (both DurableTask and Http), and DurableTask Generators. These are vital to enhance the functionality and performance of the project.
Additionally, a new build target 'CopyGrpcNativeAssetsToOutDir' is introduced to ensure that GRPC native assets are copied to the output directory for OSX platform. This support is crucial for running the project on OSX environment.