feat: add Azure Subscription Id to settings and update VM listing method
In line with the objective to be able to fetch and ascertain the number of Virtual Machines (VMs) linked to a given Azure subscription, this commit introduces the "AZ_SUBSCRIPTION_ID" to the local settings, and updates the method for listing Azure VMs.

A new parameter, the subscription Id, is now passed to the 'ListAzureVirtualMachinesAsync' method, which uses the Azure ResourceManager.Compute to fetch the VMs related to this Id. The method now returns a list of Virtual Machine Resources instead of performing a mere log. This fetch operation is performed via 'CallActivityAsync', and the method has been adjusted accordingly. Unused using directives have also been cleaned up.